requirements.txt
pip~=21.1.3
setuptools~=56.0.0
asgiref==3.4.1
Django==3.2.9
django-filter==21.1
djangorestframework==3.12.4
et-xmlfile==1.1.0
openpyxl==3.0.9
pytz==2021.3
sqlparse==0.4.2
xlrd==2.0.1
